feat: 1

+++ src/main/java/com/example/survey/repository/VisitFrequencyOptionsRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.survey.entity.AgeGroupsEntity;
 import com.example.survey.entity.VisitFrequencyOptionsEntity;
 
-public interface VisitFrequencyOptionsRepository extends JpaRepository<VisitFrequencyOptionsEntity, Long>{
+public interface VisitFrequencyOptionsRepository extends JpaRepository<VisitFrequencyOptionsEntity, Integer>{
 	
 	// sort_orderの昇順で全件取得
     List<AgeGroupsEntity> findAllByOrderBySortOrderAsc();
