+++ src/main/java/com/example/survey/repository/VisitPurposeOptionsRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.survey.entity.AgeGroupsEntity;
 import com.example.survey.entity.VisitPurposeOptionsEntity;
 
-public interface VisitPurposeOptionsRepository extends JpaRepository<VisitPurposeOptionsEntity, Long>{
+public interface VisitPurposeOptionsRepository extends JpaRepository<VisitPurposeOptionsEntity, Integer>{
 	
 	// sort_orderの昇順で全件取得
     List<AgeGroupsEntity> findAllByOrderBySortOrderAsc();
